fix(demo): bldmainte / seisaku のデモシードに見積2件を追加 (#651)
kensetsu のみ見積（quotes）がシードされており、bldmainte・seisaku は見積一
覧が常に空だった。両テンプレに kensetsu と同じ構造（insertQuote ×2・
accepted/sent・document_sequences 更新）で、業種に合った品目の見積サンプル
を追加。

RecordingQueryExecutor を DemoDataSeederDateTest.php から独立ファイルへ抽出
（PSR-4 準拠・テスト間で共有可能に）し、全テンプレで見積が1件以上シードさ
れることを検証する DemoDataSeederQuoteTest.php を追加。
